backend/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.42.0
httpx==0.27.2
PyYAML==6.0.2
faiss-cpu==1.8.0.post1
python-docx==1.1.2
pypdf==4.3.1
numpy==1.26.4
tiktoken==0.7.0